refacto(oktaApiAuth/test): split oktaApiAuth tests into multiple files
