+++ pkg/oktaApiAuth/oktaApiAuth_test.go
@@ -1,11 +1,9 @@
 package oktaApiAuth
 
 import (
-	"crypto/tls"
 	"fmt"
 	"net/http"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/h2non/gock.v1"
@@ -43,315 +41,6 @@ echo -n | openssl s_client -connect example.oktapreview.com:443 2>/dev/null |\
  openssl dgst -sha256 -binary | base64
 */
 var pin []string = []string{"SE4qe2vdD9tAegPwO79rMnZyhHvqj3i5g1c2HkyGUNE="}
-
-type authTest struct {
-	testName      string
-	mfaRequired   bool
-	passcode      string
-	requests      []authRequest
-	unmatchedReq  bool
-	allowedGroups string
-	err           error
-}
-
-type setupTest struct {
-	testName string
-	requests []authRequest
-	err      error
-}
-
-type authRequest struct {
-	path             string
-	payload          map[string]string
-	httpStatus       int
-	jsonResponseFile string
-}
-
-func TestOktaReq(t *testing.T) {
-	defer gock.Off()
-	// Uncomment the following line to see HTTP requests intercepted by gock
-	//gock.Observe(gock.DumpRequest)
-
-	tests := []setupTest{
-		{
-			"invalid json response - failure",
-			[]authRequest{
-				{
-					"/api/v1/authn",
-					map[string]string{"username": username, "password": password},
-					http.StatusOK,
-					"invalid.json",
-				},
-			},
-			fmt.Errorf("invalid character '-' looking for beginning of object key string"),
-		},
-		{
-			"invalid payload - failure",
-			[]authRequest{
-				{
-					"/api/v1/authn",
-					nil,
-					http.StatusOK,
-					"invalid.json",
-				},
-			},
-			fmt.Errorf("invalid character '-' looking for beginning of object key string"),
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			gock.Clean()
-			gock.Flush()
-
-			apiCfg := &OktaAPIConfig{
-				Url:                 oktaEndpoint,
-				Token:               token,
-				UsernameSuffix:      "algolia.com",
-				AssertPin:           pin,
-				MFARequired:         false,
-				AllowUntrustedUsers: true,
-				MFAPushMaxRetries:   20,
-				MFAPushDelaySeconds: 3,
-			}
-			userCfg := &OktaUserConfig{
-				Username: username,
-				Password: password,
-				Passcode: "",
-				ClientIp: ip,
-			}
-
-			for _, req := range test.requests {
-				reqponseFile := fmt.Sprintf("../../testing/fixtures/oktaApi/%s", req.jsonResponseFile)
-				l := gock.New(oktaEndpoint)
-				l = l.Post(req.path).
-					MatchHeader("Authorization", fmt.Sprintf("SSWS %s", token)).
-					MatchHeader("X-Forwarded-For", ip).
-					MatchType("json").
-					JSON(req.payload)
-				l.Reply(req.httpStatus).
-					File(reqponseFile)
-			}
-
-			a := NewOktaApiAuth()
-			assert.NotNil(t, a)
-			a.ApiConfig = apiCfg
-			a.UserConfig = userCfg
-			err := a.InitPool()
-			assert.Nil(t, err)
-			gock.InterceptClient(a.pool)
-			// Lets ensure we wont reach the real okta API
-			gock.DisableNetworking()
-			_, err = a.oktaReq(http.MethodPost, test.requests[0].path, test.requests[0].payload)
-			if test.err == nil {
-				assert.Nil(t, err)
-			} else {
-				assert.Equal(t, test.err.Error(), err.Error())
-			}
-		})
-	}
-}
-
-type allowedGroupsTest struct {
-	testName      string
-	requests      []authRequest
-	allowedGroups string
-	err           error
-}
-
-func TestCheckAllowedGroups(t *testing.T) {
-	defer gock.Off()
-	// Uncomment the following line to see HTTP requests intercepted by gock
-	//gock.Observe(gock.DumpRequest)
-
-	tests := []allowedGroupsTest{
-		{
-			"invalid json response - failure",
-			[]authRequest{
-				{
-					fmt.Sprintf("/api/v1/users/%s/groups", username),
-					map[string]string{"username": username, "password": password},
-					http.StatusOK,
-					"invalid_groups.json",
-				},
-			},
-			"test1, test2",
-			fmt.Errorf("invalid character '-' in numeric literal"),
-		},
-		{
-			"Member of AllowedGroups - success",
-			[]authRequest{
-				{
-					fmt.Sprintf("/api/v1/users/%s/groups", username),
-					map[string]string{"username": username, "password": password},
-					http.StatusOK,
-					"groups.json",
-				},
-			},
-			"Cloud App Users, test2",
-			nil,
-		},
-		{
-			"Not member of AllowedGroups - failure",
-			[]authRequest{
-				{
-					fmt.Sprintf("/api/v1/users/%s/groups", username),
-					map[string]string{"username": username, "password": password},
-					http.StatusOK,
-					"groups.json",
-				},
-			},
-			"test1, test2",
-			fmt.Errorf("Not mmember of an AllowedGroup"),
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			gock.Clean()
-			gock.Flush()
-
-			apiCfg := &OktaAPIConfig{
-				Url:                 oktaEndpoint,
-				Token:               token,
-				UsernameSuffix:      "algolia.com",
-				AssertPin:           pin,
-				MFARequired:         false,
-				AllowUntrustedUsers: true,
-				MFAPushMaxRetries:   20,
-				MFAPushDelaySeconds: 3,
-				AllowedGroups:       test.allowedGroups,
-			}
-			userCfg := &OktaUserConfig{
-				Username: username,
-				Password: password,
-				Passcode: "",
-				ClientIp: ip,
-			}
-
-			for _, req := range test.requests {
-				reqponseFile := fmt.Sprintf("../../testing/fixtures/oktaApi/%s", req.jsonResponseFile)
-				l := gock.New(oktaEndpoint)
-				l = l.Get(req.path).
-					MatchHeader("Authorization", fmt.Sprintf("SSWS %s", token)).
-					MatchHeader("X-Forwarded-For", ip).
-					MatchType("json")
-				l.Reply(req.httpStatus).
-					File(reqponseFile)
-			}
-
-			a := NewOktaApiAuth()
-			assert.NotNil(t, a)
-			a.ApiConfig = apiCfg
-			a.UserConfig = userCfg
-			err := a.InitPool()
-			assert.Nil(t, err)
-			gock.InterceptClient(a.pool)
-			// Lets ensure we wont reach the real okta API
-			gock.DisableNetworking()
-			err = a.checkAllowedGroups()
-			if test.err == nil {
-				assert.Nil(t, err)
-			} else {
-				assert.Equal(t, test.err.Error(), err.Error())
-			}
-		})
-	}
-}
-
-type poolTest struct {
-	testName string
-	host     string
-	port     string
-	pinset   []string
-	err      error
-}
-
-func startTLS(t *testing.T) {
-	t.Helper()
-
-	mux := http.NewServeMux()
-	mux.HandleFunc("/", func(w http.ResponseWriter, req *http.Request) {
-		w.Header().Set("Content-Type", "text/plain; charset=utf-8")
-		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write([]byte("This is an example server.\n"))
-	})
-	cfg := &tls.Config{MinVersion: tls.VersionTLS12}
-	s := http.Server{
-		Addr:         fmt.Sprintf("%s:%s", tlsHost, tlsPort),
-		Handler:      mux,
-		TLSConfig:    cfg,
-		ReadTimeout:  1 * time.Second,
-		WriteTimeout: 1 * time.Second,
-	}
-	err := s.ListenAndServeTLS("../../testing/fixtures/utils/server.crt",
-		"../../testing/fixtures/utils/server.key")
-	assert.NoError(t, err)
-	t.Cleanup(func() { _ = s.Close() })
-}
-
-func TestInitPool(t *testing.T) {
-	invalidHost := "invalid{host"
-	invalidHostErr := fmt.Sprintf("parse \"https://%s:%s\": invalid character \"{\" in host name",
-		invalidHost,
-		tlsPort)
-
-	tests := []poolTest{
-		{
-			"Test valid pinset",
-			tlsHost,
-			tlsPort,
-			[]string{validPinset},
-			nil,
-		},
-
-		{
-			"Test invalid pinset",
-			tlsHost,
-			tlsPort,
-			[]string{invalidPinset},
-			fmt.Errorf("Server pubkey does not match pinned keys"),
-		},
-
-		{
-			"Test unreachable host",
-			tlsHost,
-			"1444",
-			[]string{},
-			fmt.Errorf(fmt.Sprintf("dial tcp %s:1444: connect: connection refused", tlsHost)),
-		},
-
-		{
-			"Test invalid url",
-			invalidHost,
-			tlsPort,
-			[]string{},
-			fmt.Errorf(invalidHostErr),
-		},
-	}
-
-	go func() {
-		startTLS(t)
-	}()
-
-	time.Sleep(1 * time.Second)
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			a := NewOktaApiAuth()
-			a.ApiConfig.Url = fmt.Sprintf("https://%s:%s", test.host, test.port)
-			a.ApiConfig.AssertPin = test.pinset
-			err := a.InitPool()
-			if test.err == nil {
-				if err != nil {
-					t.Logf(err.Error())
-				}
-				assert.Nil(t, err)
-			} else {
-				assert.Equal(t, test.err.Error(), err.Error())
-			}
-		})
-	}
-}
 
 func TestAuth(t *testing.T) {
 	defer gock.Off()
